# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BudgetFriendly API
  version: 1.0.0
servers:
  - url: https://api.example.com
  - url: http://localhost:3001

security:
  - bearerAuth: []

tags:
  - name: Auth
  - name: Users
  - name: Transactions
  - name: Budgets
  - name: Goals
  - name: Analytics

paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, refreshToken]
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '204':
          description: Logged out
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Email a 6-digit password reset code
      description: Always returns 200 so the endpoint can't be used to discover accounts. Requests within 60 seconds of a previous code are ignored.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
                  devCode:
                    type: string
                    description: Only present when AUTH_DEV_EXPOSE_RESET_CODE=1 (local testing).
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with an emailed code and sign in
      description: Codes expire after PASSWORD_RESET_TTL_MIN minutes (default 15) and allow 5 attempts. A successful reset revokes every other session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, newPassword]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^\d{6}$' }
                newPassword: { type: string, minLength: 8 }
                deviceName: { type: string, maxLength: 100 }
      responses:
        '200':
          description: Password changed; authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '204':
          description: Password changed
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/auth/sessions:
    get:
      tags: [Auth]
      summary: Signed-in devices for this account (current one flagged)
      responses:
        '200': { description: OK }
  /v1/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Sign out one device (use id "revoke-others" with POST to sign out every other device)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
    post:
      tags: [Auth]
      summary: POST /v1/auth/sessions/revoke-others signs out all other devices
      parameters: [{ name: id, in: path, required: true, schema: { type: string, enum: [revoke-others] } }]
      responses:
        '200': { description: OK }

  /v1/recurring:
    get:
      tags: [Recurring]
      summary: Recurring transactions and bills
      responses:
        '200': { description: OK }
    post:
      tags: [Recurring]
      summary: Create a schedule (type, amount, category, frequency weekly|monthly|yearly, startDate, autoCreate, remindDaysBefore, budgetCategory)
      responses:
        '201': { description: Created; includes how many due occurrences were recorded }
  /v1/recurring/{id}:
    patch:
      tags: [Recurring]
      summary: Update or pause a schedule
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
    delete:
      tags: [Recurring]
      summary: Delete a schedule (already recorded transactions stay)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deleted }
  /v1/recurring/run:
    post:
      tags: [Recurring]
      summary: Record everything due for the signed-in user
      responses:
        '200': { description: OK }

  /v1/cron/daily:
    get:
      tags: [Jobs]
      summary: Vercel Cron job — recurring transactions, bill reminders, bank sync. Requires Authorization Bearer CRON_SECRET.
      security: []
      responses:
        '200': { description: OK }

  /v1/notifications:
    get:
      tags: [Notifications]
      summary: In-app notification feed (newest first) with unread count; ?before=ISO for older
      responses:
        '200': { description: OK }
  /v1/notifications/read:
    post:
      tags: [Notifications]
      summary: Mark notifications read (body ids[], or all when omitted)
      responses:
        '200': { description: OK }
  /v1/notifications/prefs:
    get:
      tags: [Notifications]
      summary: Delivery preferences (paceAlerts, billReminders, weeklyCheckIn, autoSave)
      responses:
        '200': { description: OK }
    patch:
      tags: [Notifications]
      summary: Update delivery preferences
      responses:
        '200': { description: OK }
  /v1/push-tokens:
    post:
      tags: [Notifications]
      summary: Register this device's Expo push token
      responses:
        '200': { description: OK }
    delete:
      tags: [Notifications]
      summary: Unregister a push token (on log out)
      responses:
        '200': { description: OK }

  /v1/budgets/{id}/rollover:
    get:
      tags: [Budgets]
      summary: Unspent money in an ended budget and where it can go
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
    post:
      tags: [Budgets]
      summary: Move unspent money into a goal or the next budget's Savings bucket (once per budget)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '409': { description: Already rolled over }
  /v1/budgets/{id}/members:
    get:
      tags: [Budgets]
      summary: Owner and members of a shared budget
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
  /v1/budgets/{id}/members/{userId}:
    delete:
      tags: [Budgets]
      summary: Owner removes a member, or a member leaves
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /v1/budgets/{id}/invites:
    post:
      tags: [Budgets]
      summary: Owner creates a one-time 6-character invite code (valid 7 days)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '201': { description: Created }
  /v1/budget-invites/accept:
    post:
      tags: [Budgets]
      summary: Join a shared budget with an invite code
      responses:
        '200': { description: OK }

  /v1/bank-links/mono:
    post:
      tags: [Bank links]
      summary: Finish Mono Connect — exchange the code, save the account, import the last 90 days for review
      responses:
        '201': { description: Connected }
        '501': { description: MONO_SECRET_KEY not configured }
  /v1/bank-links/{id}/sync:
    post:
      tags: [Bank links]
      summary: Import new transactions from a live connection now
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '409': { description: The bank needs to be reconnected }

  /v1/users/me:
    get:
      tags: [Users]
      summary: Get profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Users]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      parameters:
        - in: query
          name: start
          schema: { type: string, format: date }
        - in: query
          name: end
          schema: { type: string, format: date }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string, enum: [income, expense] }
        - in: query
          name: category
          schema: { type: string }
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Transaction' }
                  nextCursor:
                    type: string
                    nullable: true
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Transactions]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transaction
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Transactions]
      summary: Update transaction
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Transactions]
      summary: Delete transaction
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /v1/budgets:
    get:
      tags: [Budgets]
      summary: List budgets
      parameters:
        - in: query
          name: start
          schema: { type: string, format: date }
        - in: query
          name: end
          schema: { type: string, format: date }
      responses:
        '200':
          description: Budgets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Budget' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Budgets]
      summary: Create budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [budget]
                properties:
                  budget:
                    $ref: '#/components/schemas/Budget'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/budgets/{id}:
    get:
      tags: [Budgets]
      summary: Get budget
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Budget
          content:
            application/json:
              schema:
                type: object
                required: [budget]
                properties:
                  budget: { $ref: '#/components/schemas/Budget' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Budgets]
      summary: Update budget
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBudgetRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [budget]
                properties:
                  budget: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Budgets]
      summary: Delete budget
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /v1/goals:
    get:
      tags: [Goals]
      summary: List goals
      responses:
        '200':
          description: Goals
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Goal' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Goals]
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [goal]
                properties:
                  goal: { $ref: '#/components/schemas/Goal' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }

  /v1/goals/{id}:
    patch:
      tags: [Goals]
      summary: Update goal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [goal]
                properties:
                  goal: { $ref: '#/components/schemas/Goal' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Goals]
      summary: Delete goal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }

  /v1/analytics/summary:
    get:
      tags: [Analytics]
      summary: Summary for a date range
      parameters:
        - in: query
          name: start
          required: true
          schema: { type: string, format: date }
        - in: query
          name: end
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
        '401': { $ref: '#/components/responses/ErrorResponse' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true

    User:
      type: object
      required: [id, email, createdAt, updatedAt]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        currency: { type: string, nullable: true }
        locale: { type: string, nullable: true }
        monthlyIncome: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        deviceName: { type: string }

    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user: { $ref: '#/components/schemas/User' }
        accessToken: { type: string }
        refreshToken: { type: string }

    Transaction:
      type: object
      required: [id, type, amount, category, description, occurredAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        category: { type: string }
        description: { type: string }
        occurredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    CreateTransactionRequest:
      type: object
      required: [type, amount, category, description, occurredAt]
      properties:
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        category: { type: string }
        description: { type: string }
        occurredAt: { type: string, format: date-time }

    UpdateTransactionRequest:
      type: object
      properties:
        type: { type: string, enum: [income, expense] }
        amount: { type: number }
        category: { type: string }
        description: { type: string }
        occurredAt: { type: string, format: date-time }

    BudgetCategory:
      type: object
      required: [name, budgeted]
      properties:
        name: { type: string }
        budgeted: { type: number }
        spent: { type: number, nullable: true }

    Budget:
      type: object
      required: [id, name, totalBudget, period, startDate, categories, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        totalBudget: { type: number }
        period: { type: string, enum: [monthly, weekly] }
        startDate: { type: string, format: date }
        categories:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/BudgetCategory'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    CreateBudgetRequest:
      type: object
      required: [name, totalBudget, period, startDate, categories]
      properties:
        name: { type: string }
        totalBudget: { type: number }
        period: { type: string, enum: [monthly, weekly] }
        startDate: { type: string, format: date }
        categories:
          type: object
          additionalProperties:
            type: object
            required: [budgeted]
            properties:
              budgeted: { type: number }

    UpdateBudgetRequest:
      type: object
      properties:
        name: { type: string }
        totalBudget: { type: number }
        period: { type: string, enum: [monthly, weekly] }
        startDate: { type: string, format: date }
        categories:
          type: object
          additionalProperties:
            type: object
            properties:
              budgeted: { type: number }

    Goal:
      type: object
      required: [id, name, targetAmount, currentAmount, targetDate, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        targetAmount: { type: number }
        currentAmount: { type: number }
        targetDate: { type: string, format: date }
        emoji: { type: string, nullable: true }
        color: { type: string, nullable: true }
        category: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    CreateGoalRequest:
      type: object
      required: [name, targetAmount, targetDate]
      properties:
        name: { type: string }
        targetAmount: { type: number }
        currentAmount: { type: number, default: 0 }
        targetDate: { type: string, format: date }
        emoji: { type: string }
        color: { type: string }
        category: { type: string }

    UpdateGoalRequest:
      type: object
      properties:
        name: { type: string }
        targetAmount: { type: number }
        currentAmount: { type: number }
        targetDate: { type: string, format: date }
        emoji: { type: string }
        color: { type: string }
        category: { type: string }

    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        currency: { type: string }
        locale: { type: string }
        monthlyIncome: { type: number }

    AnalyticsSummary:
      type: object
      required: [totalBalance, income, expenses, remainingBudget, spendingByCategory]
      properties:
        totalBalance: { type: number }
        income: { type: number }
        expenses: { type: number }
        remainingBudget: { type: number }
        spendingByCategory:
          type: object
          additionalProperties: { type: number }